Adds the UserProfileUpdateSerializer, userprofile view,
